j'ai modifie le type des id dans MicroSEmployee

+++ microservice-employe/src/main/java/com/employe/microemploye/dao/ServiceDao.java
@@ -1,14 +1,11 @@
 package com.employe.microemploye.dao;
 
-import com.employe.microemploye.model.Employee;
 import com.employe.microemploye.model.Service;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
-public interface ServiceDao extends JpaRepository<Service, Integer> {
+public interface ServiceDao extends JpaRepository<Service, Long> {
 
 }
